[#5] Rename OptimalVector to Map
