+++ benchmark/OptimalVector.hs
@@ -25,7 +25,7 @@ import Data.Proxy (Proxy (..))
 import Data.Typeable (Typeable)
 import GHC.TypeLits
 
-import Data.TypeRep.OptimalVector
+import Data.TypeRep.Map (TF (..), TypeRepMap (..), fromList, lookup)
 
 benchVectorOpt :: Benchmark
 benchVectorOpt = bgroup "vector optimal"
@@ -34,7 +34,7 @@ benchVectorOpt = bgroup "vector optimal"
    -- , bench "update old" $ whnf (\x -> rknf $ insert x bigMap) (Proxy :: Proxy 1)
    ]
 
-tenLookups :: TypeRepVector (Proxy :: Nat -> *)
+tenLookups :: TypeRepMap (Proxy :: Nat -> *)
            -> ( Proxy 10, Proxy 20, Proxy 30, Proxy 40
               , Proxy 50, Proxy 60, Proxy 70, Proxy 80
               )
@@ -44,14 +44,14 @@ tenLookups tmap = (lp, lp, lp, lp, lp, lp, lp, lp)
     lp = fromJust $ lookup tmap
 
 -- TypeRepMap of 10000 elements
-bigMap :: TypeRepVector (Proxy :: Nat -> *)
+bigMap :: TypeRepMap (Proxy :: Nat -> *)
 bigMap = fromList $ buildBigMap 10000 (Proxy :: Proxy 0) []
 
 buildBigMap :: forall a . (KnownNat a) => Int -> Proxy (a :: Nat) -> [TF (Proxy :: Nat -> *)] -> [TF (Proxy :: Nat -> *)]
 buildBigMap 1 x = (TF x :)
 buildBigMap n x = (TF x :) . buildBigMap (n - 1) (Proxy :: Proxy (a + 1))
 
-rknf :: TypeRepVector f -> ()
+rknf :: TypeRepMap f -> ()
 rknf tVect = rnf (fingerprintAs tVect, fingerprintBs tVect)
 
 prepareBenchVectorOpt :: IO ()
